[otel-contrib] Address comments.
* Fix docs.

* Move composite propagator tests back and use dummy propagator to test it.
